separate videodit from dit
